// src/shared/api/reabo/generated/types.gen.ts
export type PaiementDetails = {
    payer?: string | null;
    country?: string | null;
    readonly service?: string | null;
    currency?: string | null;
    message?: string | null;
    fees?: boolean;
    amount?: number;
    methodePaiement?: number;
    reference?: string | null;
};
// ...
export type ReaboRequest = {
    offreCode: string;
    option?: string | null;
    subscriptionNumber: number;
    duree: number;
    telephone?: string | null;
    numabo: string;
    codePromo?: string | null;
    paiementDetails: PaiementDetails;
};
// ...
export type UpgrateRequest = {
    numabo: number;
    offreCode: string;
    subscriptionNumber: number;
    option?: Array<string> | null;
    codePromo?: string | null;
    paiementDetails?: PaiementDetails;
};